@@ -1,5 +1,6 @@
 import 'package:sqflite_common/sqlite_api.dart';
 import 'package:sqflite_common/src/constant.dart';
+import 'package:sqflite_common/src/factory_mixin.dart';
 import 'package:sqflite_common/src/mixin/factory.dart';
 import 'package:test/test.dart';
 
@@ -16,11 +17,14 @@ void main() {
     test('buildDatabaseFactory', () async {
       final methods = <String>[];
       final factory = buildDatabaseFactory(
+          tag: 'mock',
           invokeMethod: (String method, [dynamic arguments]) async {
-        final dynamic result = mockResult(method);
-        methods.add(method);
-        return result;
-      });
+            final dynamic result = mockResult(method);
+            methods.add(method);
+            return result;
+          });
+      expect((factory as SqfliteDatabaseFactoryMixin).tag, 'mock');
+      // ignore: unnecessary_type_check
       expect(factory is SqfliteInvokeHandler, isTrue);
       await factory.openDatabase(inMemoryDatabasePath);
       expect(methods, <String>['openDatabase']);
